+++ mopub-sdk/src/main/java/com/mopub/common/MoPub.java
@@ -8,7 +8,7 @@ import com.mopub.mobileads.MoPubRewardedVideoListener;
 import com.mopub.mobileads.MoPubRewardedVideoManager;
 
 public class MoPub {
-    public static final String SDK_VERSION = "3.10.0";
+    public static final String SDK_VERSION = "3.11.0";
 
     public static enum LocationAwareness { NORMAL, TRUNCATED, DISABLED }